Extract git-actions region into sections/git-actions.js
Fourth of P-08's six region extractions. openPath / openInTerminal / workOn / openFolder all move; the three callers into the terminal subsystem (_termCreateTab + _termSyncOpenFlag via tab-drag.js, _termActiveTab + _termSendResize + termState via dashboard-main.js) now ride their respective module boundaries.
