+++ frontend/src/js/dashboard-main.js
@@ -30,6 +30,9 @@ import {
 import {
     _refreshShadowCache, _consumeShadowCache,
 } from './sections/shadow-cache.js';
+import {
+    openPath, openInTerminal, workOn, openFolder,
+} from './sections/git-actions.js';
 
 /* --- In-app config editor --- */
 function _setField(form, name, value) {
@@ -447,128 +450,6 @@ async function saveConfig(ev) {
     } catch (e) {
         errEl.textContent = 'Save failed: ' + e;
         errEl.style.display = 'block';
-    }
-}
-
-/* --- Git row actions: open in IDEA / VS Code / terminal --- */
-async function openPath(ev, path, tool) {
-    ev.stopPropagation();
-    var btn = ev.currentTarget;
-    // Buttons now hold an inline SVG, so save/restore innerHTML.
-    var originalHtml = btn.innerHTML;
-    btn.disabled = true;
-    function restore() {
-        btn.innerHTML = originalHtml;
-        btn.classList.remove('is-ok', 'is-err');
-        btn.disabled = false;
-    }
-    function flash(cls, label, ms) {
-        btn.classList.add(cls);
-        btn.textContent = label;
-        setTimeout(restore, ms);
-    }
-    try {
-        var res = await fetch('/open', {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json'},
-            body: JSON.stringify({path: path, tool: tool})
-        });
-        if (!res.ok) {
-            flash('is-err', 'err', 1200);
-            return;
-        }
-        flash('is-ok', 'ok', 800);
-    } catch (e) {
-        flash('is-err', 'err', 1200);
-    }
-}
-
-/* Open the integrated terminal pane and spawn a fresh tab cwd'd at the
-   given path. The button lives next to the "open with" slots on every
-   git row; path is the same absolute directory those slots use, and is
-   re-validated server-side against the workspace/worktrees sandbox
-   before the pty is forked. */
-function openInTerminal(ev, path) {
-    if (ev) { ev.stopPropagation(); ev.preventDefault(); }
-    var pane = document.getElementById('term-pane');
-    if (pane.hasAttribute('hidden')) {
-        pane.removeAttribute('hidden');
-        _termSyncOpenFlag(true);
-        localStorage.setItem('term-open', '1');
-    }
-    var side = termState.lastFocused === 'right' ? 'right' : 'left';
-    // Default tab label to the target directory basename so several tabs
-    // opened from different repos are visually distinct. Writes to
-    // customName, so double-click-to-rename still overrides it and the
-    // choice persists across reloads via _termPersistTabs.
-    var basename = String(path).replace(/\/+$/, '').split('/').pop() || '';
-    _termCreateTab(side, {cwd: path, customName: basename});
-    setTimeout(function() {
-        var tab = _termActiveTab();
-        if (tab) { _termSendResize(tab); tab.term.focus(); }
-    }, 0);
-}
-
-/* Per-card "work on <slug>" button — inject the text at the prompt of
-   the currently focused terminal tab, no auto-Enter (user confirms). If
-   no tab is open, open the terminal pane, spawn one, and inject once
-   its WebSocket is ready. */
-function workOn(ev, slug) {
-    if (ev) { ev.stopPropagation(); ev.preventDefault(); }
-    var text = 'work on ' + slug;
-    var active = _termActiveTab();
-    if (active && active.ws && active.ws.readyState === WebSocket.OPEN) {
-        active.ws.send(new TextEncoder().encode(text));
-        active.term.focus();
-        return;
-    }
-    var pane = document.getElementById('term-pane');
-    if (pane.hasAttribute('hidden')) {
-        pane.removeAttribute('hidden');
-        _termSyncOpenFlag(true);
-        localStorage.setItem('term-open', '1');
-    }
-    if (termState.tabs.length === 0) _termCreateTab('left');
-    // Fresh tab: ws.open is async — poll briefly until the socket is ready.
-    var tries = 0;
-    (function trySend() {
-        var tab = _termActiveTab();
-        if (tab && tab.ws && tab.ws.readyState === WebSocket.OPEN) {
-            tab.ws.send(new TextEncoder().encode(text));
-            tab.term.focus();
-            return;
-        }
-        if (++tries < 40) setTimeout(trySend, 75);
-    })();
-}
-
-/* Per-card "open folder" button — hand the item's folder to the OS
-   default file manager. Mirrors openPath's transient ok/err feedback. */
-async function openFolder(ev, relPath) {
-    if (ev) { ev.stopPropagation(); ev.preventDefault(); }
-    var btn = ev.currentTarget;
-    var originalHtml = btn.innerHTML;
-    btn.disabled = true;
-    function restore() {
-        btn.innerHTML = originalHtml;
-        btn.classList.remove('is-ok', 'is-err');
-        btn.disabled = false;
-    }
-    function flash(cls, label, ms) {
-        btn.classList.add(cls);
-        btn.textContent = label;
-        setTimeout(restore, ms);
-    }
-    try {
-        var res = await fetch('/open-folder', {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json'},
-            body: JSON.stringify({path: relPath})
-        });
-        if (!res.ok) { flash('is-err', 'err', 1200); return; }
-        flash('is-ok', 'ok', 800);
-    } catch (e) {
-        flash('is-err', 'err', 1200);
     }
 }
